Extract helper classes
