+++ FusionFall-Mod/MainWindow.axaml.cs
@@ -7,191 +7,11 @@ using System.Threading.Tasks;
 using Avalonia.Controls;
 using Avalonia.Interactivity;
 using MsBox.Avalonia;
+using FusionFall_Mod.Models;
+using FusionFall_Mod.Utilities;
 
 namespace FusionFall_Mod
 {
-    /// <summary>
-    /// Класс, представляющий заголовок Unity-файла.
-    /// </summary>
-    public class UnityHeader
-    {
-        // Константы, описывающие размеры и флаги.
-        public const int MainHeaderSize = 64;
-        public const int DataStartOffset = 512;
-        public const int InitialDataSize = 513;
-
-        public const string DefaultFlag = "UnityWeb";
-        public const string RetroFlag = "streamed";
-
-        /// <summary>
-        /// Флаг, определяющий тип файла. По умолчанию - DefaultFlag.
-        /// </summary>
-        public string FlagFile { get; }
-
-        /// <summary>
-        /// Основная (мажорная) версия.
-        /// </summary>
-        public byte MajorVersion { get; set; } = 2;
-
-        /// <summary>
-        /// Дополнительная информация о версии.
-        /// </summary>
-        public string VersionInfo { get; set; } = "fusion-2.x.x";
-
-        /// <summary>
-        /// Инфо о сборке .
-        /// </summary>
-        public string BuildInfo { get; set; } = "2.5.4b5";
-
-        /// <summary>
-        /// Полный размер файла.
-        /// </summary>
-        public int FileSize { get; set; } = 0;
-
-        /// <summary>
-        /// Смещение начала данных (по умолчанию указывает на MainHeaderSize).
-        /// </summary>
-        public int FirstOffset { get; set; } = MainHeaderSize;
-
-        /// <summary>
-        /// Размер сжатых данных.
-        /// </summary>
-        public int FileZipSize { get; set; } = 0;
-
-        /// <summary>
-        /// Размер распакованных данных (по умолчанию InitialDataSize).
-        /// </summary>
-        public int FileUnzipSize { get; set; } = InitialDataSize;
-
-        /// <summary>
-        /// Смещение конца данных (или конец файла).
-        /// </summary>
-        public int LastOffset { get; set; } = 0;
-
-        /// <summary>
-        /// Конструктор, принимающий флаг файла.
-        /// </summary>
-        /// <param name="flag">
-        /// Допустимые значения: DefaultFlag ("UnityWeb") или RetroFlag ("streamed").
-        /// Если строка не совпадает ни с одним из них, автоматически устанавливается DefaultFlag.
-        /// </param>
-        public UnityHeader(string flag)
-        {
-            if (string.Equals(flag, DefaultFlag, StringComparison.OrdinalIgnoreCase) ||
-                string.Equals(flag, RetroFlag, StringComparison.OrdinalIgnoreCase))
-            {
-                FlagFile = flag;
-            }
-            else
-            {
-                FlagFile = DefaultFlag;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Утилитный класс для конвертации целых чисел в Big-Endian и обратно.
-    /// </summary>
-    public static class EndianConverter
-    {
-        /// <summary>
-        /// Преобразует int в массив байт Big-Endian.
-        /// </summary>
-        public static byte[] ToBigEndian(int value)
-        {
-            byte[] bytes = BitConverter.GetBytes(value);
-            if (BitConverter.IsLittleEndian)
-                Array.Reverse(bytes);
-            return bytes;
-        }
-
-        /// <summary>
-        /// Читает int из массива байт в Big-Endian формате, начиная с указанного индекса.
-        /// </summary>
-        public static int FromBigEndian(byte[] bytes, int startIndex = 0)
-        {
-            if (bytes == null)
-                throw new ArgumentNullException(nameof(bytes));
-            if (startIndex < 0 || startIndex + 4 > bytes.Length)
-                throw new ArgumentOutOfRangeException(nameof(startIndex), "Недопустимый индекс старта в массиве байт.");
-
-            // Создаем временный массив, копируем нужные 4 байта
-            byte[] temp = new byte[4];
-            Array.Copy(bytes, startIndex, temp, 0, 4);
-
-            if (BitConverter.IsLittleEndian)
-                Array.Reverse(temp);
-
-            return BitConverter.ToInt32(temp, 0);
-        }
-    }
-
-    public static class LzmaHelper
-    {
-        public static byte[] CompressData(byte[] data)
-        {
-            using (var input = new MemoryStream(data))
-            using (var output = new MemoryStream())
-            {
-                SevenZip.Compression.LZMA.Encoder encoder = new SevenZip.Compression.LZMA.Encoder();
-                SevenZip.CoderPropID[] propIDs =
-                {
-                    SevenZip.CoderPropID.LitContextBits,
-                    SevenZip.CoderPropID.LitPosBits,
-                    SevenZip.CoderPropID.PosStateBits,
-                    SevenZip.CoderPropID.DictionarySize
-                };
-                object[] properties =
-                {
-                    3,
-                    0,
-                    2,
-                    128 * 1024 * 4
-                };
-                encoder.SetCoderProperties(propIDs, properties);
-                encoder.WriteCoderProperties(output);
-                output.Write(BitConverter.GetBytes((long)data.Length), 0, 8);
-                encoder.Code(input, output, data.Length, -1, null);
-                return output.ToArray();
-            }
-        }
-
-        public static byte[] DecompressData(byte[] data)
-        {
-            using (MemoryStream input = new MemoryStream(data))
-            {
-                SevenZip.Compression.LZMA.Decoder decoder = new SevenZip.Compression.LZMA.Decoder();
-                byte[] properties = new byte[5];
-                if (input.Read(properties, 0, 5) != 5)
-                    throw new Exception("Input .lzma file is too short");
-                decoder.SetDecoderProperties(properties);
-                byte[] fileLengthBytes = new byte[8];
-                if (input.Read(fileLengthBytes, 0, 8) != 8)
-                    throw new Exception("Input .lzma file is too short");
-                long fileLength = BitConverter.ToInt64(fileLengthBytes, 0);
-                using (MemoryStream output = new MemoryStream())
-                {
-                    decoder.Code(input, output, input.Length - input.Position, fileLength, null);
-                    return output.ToArray();
-                }
-            }
-        }
-    }
-
-    public class FileEntry
-    {
-        public string FileName { get; }
-        public string FullPath { get; }
-        public long Size { get; }
-
-        public FileEntry(string fileName, string fullPath, long size)
-        {
-            FileName = fileName;
-            FullPath = fullPath;
-            Size = size;
-        }
-    }
-
     public partial class MainWindow : Window
     {
         private readonly string[] expectedFiles = new string[]
